Improper detection of non-requests module
Fixes false postive detecting the usage of the requests module
without a timeout. This resolves cases of modules with the word
"requests" in the name, but does not match the actual popular
third-party module "requests".

The fix checks the fully qualified name and ensures index 0 is
"requests". Previously, the code was match any module name with
"requests" in it.

Fixes #1010

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>