Remove markdown formatting in reStructuredText formatted README
A recent change to the README has code blocks in markdown format.
Since our README is actually reStructuredText format, this causes
syntax errors.

It's even bigger of a problem when an attempt to publish the package
with the README serving as the PyPI description. The syntax errors
prevents package publishing.

Fixes #1102

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>