Updates banner logo so it renders well in dark mode
GitHub supports a variety of themes including a dark mode. When
switching to a dark mode, the Bandit banner logo in the README
is almost unreadable. That's because the background is transparent
and the logo uses black text and black in the icon.

This change switches the logo from a transparent background to a
white one. That should resolve the issue in the various theme
modes.

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>