Add action to upload the wheel and tar-zipped distributions
We already publish to PyPI our packages. It is also useful to
publish our binaries as artifacts here on GitHub. This action
will build and publish release files as artifacts in the current
release.

It runs whenever a new release is published.

This change also adds attestation of the artifacts so that users
can verify the binary is the authentic one produced by our build.

More info on attestation here:
https://github.blog/2024-05-02-introducing-artifact-attestations-now-in-public-beta/

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>