Guard against empty call argument list (#1146)
Although probably uncommon, it is possible to pass an empty list
to one of subprocess functions. If this is done, the injection_shell
plugin raises an IndexError while checking the contents of the
list argument given.

The fix is to simply check for a non-empty list. Test case was also
added.

Fixes: #1141

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>