Removal of Python 3.8 support
As of October 1st, 2024, Python 3.8 becomes end-of-life. As such
Bandit should also no longer support the version.

This commit will bump minimum version of support to Python 3.9
and remove any legacy 3.8 specific code.

This is a breaking change for those still using Python 3.8 and
so the minor version will be bumped to 1.8.x for it.

Closes #1173

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>