Warn not error on an nonexistant test given (#1230)
When a user gives a test ID to include or skip, the current
behavior raises an exception and exits the process.

However, when tests end up getting deprecated and eventually
removed, it is a lot more user friendly to simple present
a warning to the user that the test ID given wasn't found
rather than a hard error and exit.

Fixes: #1228

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>