Metadata: rename classifier to classifiers (#1237)
Not sure if this metadata object changed names over time but some tooling such as LicenseCheck trips up if it's defined as classifier and not classifiers. 

This matches current documentation where it states classifier is an alias of classifiers according to metadata table here: https://setuptools.pypa.io/en/latest/userguide/declarative_config.html#metadata

A separate issue was opened on LicenseCheck.