Fix traceback from trojansource plugin
The trojansource plugin functions differently where it doesn't
process a file by AST node. It instead does a line by line search
for suspicious characters.

As a result, it can't rely on the linerange being automatically
set based on values fetched from the node. So it needs to set
the linerange manually.

Fixes: #1246

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>