Ensure the man page is built (#1257)
Ever since Sphinx switched to sphinx-build command to build
docs, the man page was no longer being built. This is because
there is now a separate subcommand "sphinx-build -b man" to
generate the man page file.

This change adds the separate command and ensures our publish
workflows generate the file when releasing a new bandit version.

This also adds the man page as part of the data files so it
gets included with the wheel and is installed to share/man/man1.

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>