Remove etc from list of temp paths
In the docs for B108, it lists the configurable paths it
checks for temp directory-like locations. In addition, it
includes a bullet point for "etc". This might be misinterpretted
as "/etc". I believe the original intention was to state
etcetera, but might be confusing as listed.

This change just removes it from the list to get rid of any
potential confusion.

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>