Argparse Python 3.14 enhancements
The argparse lib in Py314 now can do suggestion of an
argument when a typo is given. It only works on choices,
but regardless it can be handy to users of Bandit. Besides
Python 3.15 plans to enable this by default.

The other change is the color being disabled. Color in CLIs
defaults to True in Python 3.14. And frankly it doesn't look
good. Way too many colors being used.

https://docs.python.org/3/library/argparse.html#suggest-on-error

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>