Check whether Constant value is str
This change fixes a case of a missed check on the value
of a ast.Constant to be a str or not. PR #1323 fixed
many of these as part of the Python 3.14 compatibility
since ast.Str was removed. So when checking ast.Constant,
the value can many types of literals, not just str.

Fixes #1332

Signed-off-by: Eric Brown <eric_wade_brown@yahoo.com>